backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
openai==1.3.7
reportlab==4.0.8
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
httpx==0.25.2
jinja2==3.1.2
python-docx==1.1.0
weasyprint==60.2
html2text==2020.1.16
markdown==3.5.1
bleach==6.1.0
python-pptx==0.6.21
Pillow==10.1.0